Many changes to how classes are found and loaded.  Since as of Java 9 the method that was used to dynamically add jars and classes to the existing application classpath will no longer work, the Boot class now creates a custom classloader that is then used to dynamically load jars and classes.
Furthermore since testing with different versions of Java showed that if you use an older version of Java after having used a newer one, any custom classes that were compiled by the newer Java won't load in the older Java, and therefore will need to be re-compiled.   This change also includes that fix.