Fix Indexer exceptions when Solr id != controlnum
When indexing, if the Solr id is modified from the control
number, solrmarc would later still attempt to reference the
Solr document assuming the original controlnumber was the
Solr id.

For example with SolrMarcIndexerException.DELETE, this would
result in Solr records being mismatched and not deleted.

Also, unittests for Indexer were failing with newer Java.
Added `--add-opens=java.base/java.lang=ALL-UNNAMED` to the
and `test` target for the build to get reflection working.